Правки по MR
