+++ src/components/CarHeader/CarHeader.module.css
@@ -12,8 +12,3 @@
     border-color: #4c489d79;
     justify-content: space-between;
 }
-
-.filter{
-    height: 25px;
-    width: 25px;
-}
