+++ src/components/Layout/Layout.tsx
@@ -1,4 +1,4 @@
-import IMain from "../../../../interfaces/IMain";
+import IMain from "../../interfaces/IMain";
 import Header from "../Header/Header";
 import classes from "./Layout.module.css";
