+++ src/types/Car.ts
@@ -6,4 +6,5 @@ export type Car = {
     model: CarProperty,
     color: CarProperty
     price: number
+
 }
