+++ src/types/Filter.tsx
@@ -1,0 +1,7 @@
+import ReadonlyCar from "./ReadonlyCar";
+
+
+export type Filter = {
+    field: keyof ReadonlyCar;
+    isDesc: boolean;
+};
